Complete exercise 6
Exercise 6 (leapYears) has been completed and passed all tests. The function takes a number as input and checks if it meets the criteria for being a leap year, namely that it is divisible by 4, unless it's divisible by 100, in which case it's not a leap year, EXCEPT in cases when it's divisible by 400, in which case it is. The function uses nested if conditionals to check if the conditions are met and returns 'true' if they are and 'false' if they are not.