Complete Javascript Exercise 09
Created a program that takes in a string and checks if it is a palindrome. Starts by converting the string to lowercase, then removing all non-letter characters (this can easily be adapted to remove only all non-alphanumerical characters by changing the regex used). It takes the result of this process and creates a reversed version of it, then checks if the reverse version and the original version match. If they do, it returns true. If not, it returns false.