Complete Javascript Exercise 10
Completed Exercise 10 in the Odin Project's Javascript Exercises. This program generates a fibonacci sequence up to the index provided by the user and returns the last term of that sequence (e.g. index 3 returns term of 2, index 6 returns term of 8, and so on). It converts the user's input to a number and checks if it is positive before continuing. It then enters a simple loop adding prior terms to create the new term for as long as the user's input has not been reached. It returns the variable newTerm to represent the term at the user's chosen index.