Complete Javascript Exercise 11
Completed Exercise 11 of the Odin Project's Javascript Exercises. This exercise required that, given an array of objects with information about various books (e.g. title: 'Title', author 'Name'), you create an array just listing the titles of all the books in that first array. This program uses the .map method to extract all the 'title' items from the first array and add them to an array called 'titles', then return the 'titles' array.