Complete Javascript Exercise 12
Completed Exercise 12 of the Odin Project's Javascript Exercise. The goal of this exercise was to use array methods to take in any array of objects representing people, with properties of name, yearOfBirth, and yearOfDeath, and return the oldest person from this array. Additional tests involved edge cases such as one or more people still being alive (and therefore having no yearOfDeath). This program first checks to see if the yearOfDeath property is undefined, and if so assigns it the current year with the getFullYear method. After this, it uses a simple inequality check in a sort method to sort the people from oldest to youngest, and returns the first item in the resulting array.